backend/pyproject.toml
[tool.poetry]
name = "gptlearn"
version = "0.1.0"
description = "Django app for making flashcards with ChatGPT"
authors = ["CharlieKC <charlie.kruczko@gmail.com>"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
django-ninja = "^0.22.2"
django-extensions = "^3.2.3"
pydotplus = "^2.0.2"
django-tailwind = {extras = ["reload"], version = "^3.6.0"}
django-browser-reload = "^1.11.0"
whitenoise = "^6.5.0"
django-debug-toolbar = "^4.2.0"
django-markdownify = "^0.9.3"
pygments = "^2.16.1"
django-allauth = "^0.57.0"
openai = "^0.28.1"
django-stubs = "^4.2.4"
factory-boy = "^3.3.0"
pytest-watcher = "^0.3.4"
orjson = "^3.9.7"
uvicorn = "^0.23.2"
channels-redis = "^4.1.0"
selenium = "^4.13.0"
djangorestframework = "^3.14.0"
markdown = "^3.4.4"
django-cors-headers = "^4.2.0"
channels = "^4.0.0"
watchfiles = "^0.20.0"
django-rest-knox = "^4.2.0"
drf-spectacular = "^0.26.5"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-django = "^4.5.2"
pre-commit = "^3.4.0"
flake8 = "^6.1.0"
flake8-isort = "^6.1.0"
coverage = "^7.3.1"
pytest-sugar = "^0.9.7"
black = "^23.9.1"
djlint = "^1.34.0"
pylint-django = "^2.5.3"
mypy = "^1.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# ==== black ====
[tool.black]
line-length = 119
target-version = ['py311']


# ==== isort ====
[tool.isort]
profile = "black"
line_length = 119
known_first_party = [
    "my_awesome_project",
    "config",
]
skip = ["venv/"]
skip_glob = ["**/migrations/*.py"]


# ==== mypy ====
[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
plugins = [
    "mypy_django_plugin.main",
]

[[tool.mypy.overrides]]
# Django migrations should not produce any errors:
module = "*.migrations.*"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "flashcards.settings"


# ==== PyLint ====
[tool.pylint.MASTER]
load-plugins = [
    "pylint_django",
]
django-settings-module = "flashcards.settings"

[tool.pylint.FORMAT]
max-line-length = 119

[tool.pylint."MESSAGES CONTROL"]
disable = [
    "missing-docstring",
    "invalid-name",
]

[tool.pylint.DESIGN]
max-parents = 13

[tool.pylint.TYPECHECK]
generated-members = [
    "REQUEST",
    "acl_users",
    "aq_parent",
    "[a-zA-Z]+_set{1,2}",
    "save",
    "delete",
]


# ==== djLint ====
[tool.djlint]
blank_line_after_tag = "load,extends"
close_void_tags = true
format_css = true
format_js = true
# TODO: remove T002 when fixed https://github.com/Riverside-Healthcare/djLint/issues/687
ignore = "H006,H030,H031,T002"
include = "H017,H035"
indent = 2
max_line_length = 119
profile = "django"

[tool.djlint.css]
indent_size = 2

[tool.djlint.js]
indent_size = 2

[tool.pytest.ini_options]
addopts = "--ds=flashcards.settings"
python_files = [
    "tests.py",
    "test_*.py",
]
